nHenPi: count multiple occurrences to check for duplicates
Previously we assumed there are no dupilicates in Pi as it is unlikely
that a 6 digit sequence is valid and occurs twice in two different
locations in Pi. This code allows us to check this assumption holds as
the length of the input digits of Pi increases. Checking this assumption
is important becuase duplicate strings become a lot more feasible as
the length of Pi we look at increases