Visualiser: ensure brackets fully encapsulate key
In a prior commit a regression was introduced that meant that the
brackets for keys starting with one or more zeroes did not include the
last digits of the key. This was caused by javascript dropping leading
zeroes from the key and thus not including them in calculating the
bracket width. This patch reverts this regression, and ensures all
brackets are correctly set.