Visualiser: only load brackets when needed
This increase the efficiency of the site by only loading brackets if the
last bracket we placed down is either visible or has been scrolled past.
Changing to this less agressive approach makes the weebsite feel a lot
more responsive and allows the user to get a lot further through Pi than
before.