CoverFetcher: only pull covers down once
Previously images were pulled down every time a request was made for
them. This was slow, and meant that images could be pulled down from
nHentai multiple times if requests were made whilst the previous request
was ongoing. Note: this code does not handle the case where a cover is
requested immediately after another request has finished just finished.
However, this is a relatively unlikely event, and to handle it would
compromise the ability of the server to prune images that are accessed
infrequently. One method to handle this case would be to check if the
file we wish to create already exits in the covers directory, then only
downloading the cover if it does not already exist. This should be done
in the get_cover function.