nHenPi: export special media ID for objectionable content
Objectionable content as defined by the NZ government, primarily
containging to the guidance from the DIA and the Films, Videos, and
Publications Classification Act 1993 section 3: Meaning of
objectionable.

In the case that the content is tagged with the any of
the tags in the objectionable set, a special image should be loaded.
This has been implented as the media id being set to 'obj' so that
'obj.jpg' is loaded rather than the cover fetcher attempting to fetch
the cover. This prevents the cover fetcher from potentially downloading
objectionable content.