Visualiser: use average bracket width to size brackets
Previously, the visualiser calculated the bracket width for each bracket
on an individual basis. This does not make sense as an overwhelming
majority have the same size. Additionally the few non-standard brackets
are either the first bracket - a rare exception having an intervening
"." character - or were only a few pixels different than the other
brackets. This means it a lot of CPU time was spend calculating the
width of brackets as calculating width involves two non-trivial
operations: a complete style recalculation and page reflow.

Now instead of calculating the width of the each bracket specially, the
average of the first 24 non-special brackets (doujins 1-23 in Pi) is
used. This halves the number of style recalculations and page reflows,
increasing efficiency.