Visualiser: calculate exact position each time
Previous patches attempted to trade off accuracy for perfomance,
reducing the amount of computation for each brace placement. This lead
to a consistent, accumulating 8 px error roughly every 500 brackets,
meaning that brackets could be off by one or more digits by the end of
the page. This level of inaccuracy is not tolerable.

Since the accuracy-perfromance tradeoff was made several other changes
have been made to the site. Namely, brackets are now loaded as required,
rather than in advance, and the site now only displays the first ~20000
digits of Pi on the first page. This means the tradeoff is no longer
needed.

As such, this patch undoes the reduced computation work, and the full
accurate calculation is performed all the time, with little noticeable
performance drop.