Visualiser: display error loli on fetch failure
Previously the error loli only displayed when fetch through its own
error. This was normally due to network errors. In reality, the error
loli will not be shown in this case. It should be shown if the request
to the cover fetcher is not the OK response. This requires the fetch
callback function to manually throw the error when needed.